refactor(app): make fast refresh happy
