+++ src/api/uic-inventory/src/AuthContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const AuthContext = createContext();
