+++ src/api/uic-inventory/src/components/Pages/Home.jsx
@@ -23,7 +23,7 @@ import ky from 'ky';
 import PropTypes from 'prop-types';
 import { Fragment, useContext, useMemo, useRef } from 'react';
 import { List } from 'react-content-loader';
-import { AuthContext } from '../../AuthProvider';
+import { AuthContext } from '../../AuthContext';
 import { wellTypes } from '../../data/lookups';
 import { useOpenClosed } from '../Hooks/useOpenClosedHook';
 import { Chrome, ConfirmationModal, Header, Link, TableLoader, Tooltip, onRequestError, toast } from '../PageElements';
